requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colour==0.1.5
cryptography==42.0.5
debugpy==1.8.1
docutils==0.19
fastapi==0.109.2
fire==0.6.0
Flask==3.0.2
Flask-Cors==4.0.0
flask-sock==0.7.0
flask-talisman==1.1.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
ifaddr==0.2.0
importlib_metadata==7.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
libsass==0.23.0
markdown2==2.4.13
MarkupSafe==2.1.5
multidict==6.0.5
multipart==0.2.4
orjson==3.9.15
packaging==24.0
parso==0.8.3
progress==1.6
pscript==0.7.7
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.2
PyYAML==6.0.1
requests==2.31.0
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
termcolor==2.4.0
tqdm==4.66.2
typing_extensions==4.10.0
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.29.0
validators==0.24.0
vbuild==0.8.2
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
